requirements.txt
python==3.12.2
autograd==1.7.0
autograd-gamma==0.5.0
blessed==1.20.0
Brotli==1.1.0
ecos==2.0.14
et-xmlfile==1.1.0
feather-format==0.4.1
formulaic==1.0.2
gpustat==1.1.1
h5py==3.12.1
icare==0.1.0
inflate64==1.0.0
interface-meta==1.3.0
ipywidgets==8.1.5
joblib==1.4.2
jupyterlab_widgets==3.0.13
kmodes==0.12.2
lifelines==0.29.0
llvmlite==0.43.0
matplotlib==3.9.2
mkl-service==2.4.0
mkl_fft==1.3.10
multivolumefile==0.2.3

numba==0.60.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
openpyxl==3.1.5
osqp==0.6.7.post1
protobuf==4.25.3
psutil==6.0.0
py7zr==0.22.0
pyarrow==17.0.0
pybcj==1.0.2
pycox==0.3.0
pycryptodomex==3.20.0
pyppmd==1.1.0
PyQt5==5.15.10
PyWavelets==1.7.0
pyzstd==0.16.1
qdldl==0.1.7.post4
scikit-learn==1.5.2
scikit-survival==0.23.0
scipy==1.14.1
setuptools==72.1.0
sklearn-pandas==2.2.0
texttable==1.7.0
threadpoolctl==3.5.0
torch==2.4.1
torchaudio==2.4.0
torchtuples==0.2.2
torchvision==0.19.1
triton==3.0.0
watermark==2.5.0
wheel==0.44.0
widgetsnbextension==4.0.13
wrapt==1.16.0
